More Knife Things
Continued work on the Repeater spell. Changed over the material on the knife to a dissolvable one and changed how the knives spawn.
Rather than spawn 1 at a time when you fire, the spell instead spawns all 3 in sequence using coroutines. I also added a boolean to check whether all 3 blades are spawned, and if a blade has already been fired.